requirements.txt
beautifulsoup4==4.6.0
Flask==0.12.2
requests==2.18.4
python-dateutil==2.6.1
sklearn==0.0
scikit-learn==0.19.1
Pillow==5.0.0
numpy==1.14.0
scipy==1.0.0